removed API in service hub that internally play with data (options) in workspace rather than solution snapshot.
now, like other features, analyzer uses snapshot of options rather than mutable one from Workspace.

eventually we should only use OptionSet snapshot embeded in solution rather than one in workspace.
